requirements.txt
opencv_python==4.10.0.82
torch==2.3.1
transformers==4.42.4
sentencepiece
datasets
jiwer
tabulate
Pillow
numpy
tqdm
protobuf
evaluate
tensorboard
albumentations